requirements.txt
asgiref==3.8.1
azampay==0.5
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
djangorestframework==3.15.2
idna==3.10
phonenumbers==8.13.55
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
